format:check failure fix & output improvement
A Prettier default changed from `trailingComma: "es5"` to `trailingComma:
"all"`, and this results in a formatting change that hasn’t otherwise been
applied. We are changing the configuration for minimum change for now, but the
default change is likely correct.

Changing the workflow to help us more easily debug issues when `pnpm
format:check` fails. It will run `pnpm format` then display `git diff`. We
always `exit 1` on this alternate branch to ensure that the job still fails.

Signed-off-by: Austin Ziegler <aziegler@kineticcommerce.com>