chore: Switch Elixir tests back to pipes
As part of elixir-lang/elixir#14136, I learned about MIX_QUIET, to
suppress `Mix.shell()` output. Using this would not have worked with the
generator, but the generator was using `Mix.shell.info/1` to output the
JSON suite data. Changing that to `IO.write(:stdio, …)` means that we
can use `MIX_QUIET=1` for our tests.

Running this manually with Elixir 1.18.0 will still fail unless
`MIX_QUIET=1` is specified, but Elixir 1.18.1 or later will have this
fixed because the lock warnings are written to stderr instead of stdout.

Signed-off-by: Austin Ziegler <austin@zieglers.ca>